[#170142145] squash migrations again
